requirements.txt
Flask 
gunicorn
pytest
google-cloud-secret-manager
google-cloud-logging
google-cloud-firestore